refactor: Move ClearDataButton component to utils folder and update setupDiagram
The ClearDataButton component has been moved from the components folder to the utils folder to better organize the codebase. Additionally, the setupDiagram function in the setupDiagram.js file has been updated to remove the import of the ClearDataButton component. This change improves code maintainability and reduces clutter in the components folder.